+++ renderer/src/graphics_terminal.c
@@ -32,7 +32,7 @@ void graphics_draw_world(struct world *world)
 	UNUSED(world);
 }
 
-void graphics_draw_human(float x, float y, struct component_human *human)
+void graphics_draw_human(double x, double y, struct component_human *human)
 {
 	LOG_INFO("Entity %-3d : %s %3dy (%2.1f, %.01f)",
 			last_entity, human->gender == MALE ? "  male" : "female", human->age, x, y);
